Prover untangle contd.
